requirements.txt
discord
requests
bs4
python-dotenv